+++ packages/ethereum-indexer-cli/CHANGELOG.md
@@ -1,5 +1,12 @@
 # ethereum-indexer-cli
 
+## 0.6.14
+
+### Patch Changes
+
+- Updated dependencies
+  - ethereum-indexer-utils@0.6.9
+
 ## 0.6.13
 
 ### Patch Changes