+++ packages/ethereum-indexer-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # ethereum-indexer-utils
 
+## 0.6.9
+
+### Patch Changes
+
+- allow reading from file for deployments
+
 ## 0.6.8
 
 ### Patch Changes